Use a different store for the host.
The host was using the same store as was given in the constructor for the
underlying linker and defined memories/functions/globals.

As a result, a memory defined in a host, for example, would not get freed when
the host (and all instances) were disposed.

This prevents the use case where a module gets loaded once into a store and then
a host is created to instantiate the module.  If the operation to create the
host and instantiate the module is performed in a loop (e.g. a benchmark), then
memory continues to grow as the original store is never released.