Remove module creation methods from `Store`.
A recent change in the Wasmtime C API has made creating modules use an engine
and not a store.

This commit removes the module creation methods from `Store` and instead adds
them as static methods to `Module` which accept the engine to use.

`Store` remains as a type so that it can be used to share a store between
different hosts, if desired.

Fixes #25.