Implement externref values.
This commit implements `externref` values for .NET bindings.

Values can now be any .NET reference type.  Parameters and return values for
host functions can also be any .NET reference type and will be imported as
`externref` values.

An example was added to show passing and returning .NET strings as `externref`
values.

This commit also fixes host function callback rooting so that it is no longer
tied to the host.  Instead, the `_with_env` functions are used for creating the
function and the env is a GCHandle that is rooting the callback. A finalizer
for the env is used to free the handle.  This makes it so that the .NET
callback delegate remains rooted while the underlying Wasmtime `Func` exists.

Some refactoring was also done in `Function.cs` to streamline the code a
little.