Implement support for function references.
This commit implements support for function references in the .NET API.

A `Function` can now be created with `Function.FromCallback` and passed as a
value to WebAssembly; this will pass the value as a `funcref`.

A `Function` can also be used as a parameter for a host function that accepts a
`funcref`.  The new method `Invoke` can be used to call the function reference.

This also adds a new example project showing some simple function reference
interactions.