Implement module linking.
This commit implements the changes to support the module linking proposal.

Modules and instances are now supported import types for module instantiation.

Modules can now also be instantiated via `Module.Instantiate` and
`ModuleExtern.Instantiate`.

There is a breaking change to the API:

* `Instance.Externs` has been removed in favor of individual properties that
  forward to the inner `Externs` instance.  Thus `Instance.Externs.Functions`
  becomes `Instance.Functions`, etc.

Additionally, the examples and the test project have been migrated to .NET 5.

Closes #45.
Fixes #40.