Reimplement the Wasmtime API for .NET.
This commit reimplements the Wasmtime API for .NET in terms of the latest
upstream changes.

Recently the Wasmtime Rust API underwent a refactoring that made Wasmtime safer
to use in multithreaded async executors.

As a result of that change, a "store context" was needed to pass through to all
store-related types. Additionally, the store-related types no longer need to
implement IDisposable because the owning store now properly handles cleanup.
This fixes a race that previously may have occurred when .NET users forgot to
call `Dispose` and cleanup was done from a finalizer.

As this was already a breaking change, I decided to remove `Host` in favor of a
`Linker` type to better encapsulate the underlying Wasmtime API. `Host` existed
before the linker implementation in Wasmtime and I have been meaning to remove
it for a while now.

Another big change is that `Instance` no longer implements `DynamicObject`.
Users should now call `Instance.GetFunction`, etc. for retrieving the exports
of an instance.

Examples, documentation, and tests were all updated to the new API.

Fixes #53.