Store the `Store` instance itself in the context data using a weak `GCHandle`, so that it can be retrieved later from the `Caller.Store` property.
Additionally, remove the `IStore` interface and change the corresponding method/constructor parameters to accept a `Store` instead.

For example, this allows to create a `funcref` value as `Function` using the associated `Store` instead of the `Caller`, so that it can be stored and then called outside of the callback.

Note: Users that previously passed a `Caller` instance to a method accepting an `IStore` will now need to use `caller.Store`.

Fixes #200