Remove slow reflection-based fallback for function callbacks.
This commit removes the slow reflection-based fallback that is performed when a
delegate is provided that does not conform to the (up to) 12 parameter and (up
to) 4 results statically checked overloads of `Function.FromCallback` and
`Linker.DefineFunction`.

As a result, functions that take more than 12 parameters or return more than 4
results must use an untyped callback where the parameters and results are
passed as spans.

The callbacks will no longer be invoked via reflection under any circumstances.