Fix the .NET API for recent Wasmtime C API changes.
This commit fixes the definition of `wasmtime_linker_define` to match what is
currently expected upstream.

As a result, items passed to `Linker.Define` must be associated with a store.

The internal `IExternal` interface was updated to expose the store associated
with the item for `LinkerDefine` to use.

Also update the target framework to 7.0 (still `netstandard2.1` as well).